sample/AS/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py38h06a4308_0
charset-normalizer=3.1.0=pypi_0
cmake=3.26.3=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
einops=0.6.0=pypi_0
filelock=3.11.0=pypi_0
fire=0.5.0=pypi_0
flatbuffers=23.3.3=pypi_0
fonttools=4.39.3=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.4.0=pypi_0
google-auth=2.17.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.54.0=pypi_0
idna=3.4=pypi_0
imageio=2.27.0=pypi_0
importlib-metadata=6.5.0=pypi_0
importlib-resources=5.12.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-lite=1.8.0=pypi_0
lightning-utilities=0.8.0=pypi_0
lit=16.0.1=pypi_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
mdurl=0.1.2=pypi_0
mediapipe=0.9.2.1=pypi_0
mmcv=2.0.0=pypi_0
mmengine=0.7.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.24.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
opencv-contrib-python=4.7.0.72=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=h7f8727e_0
packaging=23.1=pypi_0
pandas=2.0.0=pypi_0
pillow=9.5.0=pypi_0
pip=23.1=pypi_0
protobuf=3.20.3=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydeprecate=0.3.2=pypi_0
pygments=2.15.0=pypi_0
pyparsing=3.0.9=pypi_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pypi_0
pytorch-lightning=1.9.0=pypi_0
pytorch-metric-learning=2.1.0=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.3.4=pypi_0
rsa=4.9=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.9.1=pypi_0
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=pypi_0
tensorboard=2.12.2=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.2.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.0.0=pypi_0
torchmetrics=0.11.4=pypi_0
torchvision=0.15.1=pypi_0
tqdm=4.65.0=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.15=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py38h06a4308_0
xz=5.2.10=h5eee18b_1
yapf=0.32.0=pypi_0
yarl=1.8.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
